backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.8.2
pydantic-settings==2.3.3
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
